rename `secure_session` module to `ip_secure`
